Split streams out into their own files.
