+++ watcher
@@ -3,5 +3,5 @@ while : ; do
     echo "Watching for changes..."
     inotifywait -r -qq -e close_write -e create -e delete --exclude ^.+?\.sw.?$ .
     echo "File changed! Linting and rebuilding docs..."
-    eslint . && echo "Linter passed." && groc --only-render-newer
+    eslint . && echo "Linter passed." && groc --only-render-newer && mocha --recursive
 done
